refac: Todo 파일 정리
