+++ src/Components/team/todo/TodoType.ts
@@ -1,0 +1,15 @@
+export interface Todo {
+    id: number;
+    completed: boolean;
+    nickname: string;
+    createdAt: string;
+    endAt: string;
+    title: string;
+    todo: string;
+}
+
+export interface teamMember {
+    usercode: number;
+    teamId: string;
+    nickname: string;
+}
